backend/requirements.txt
absl-py==2.2.2
accelerate==1.5.2
acres==0.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.3.0
babel==2.17.0
bangla==0.0.2
banks==2.1.1
beautifulsoup4==4.13.3
bitsandbytes==0.45.4
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.37.28
botocore==1.37.28
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
contourpy==1.2.1
coqpit==0.0.17
cryptography==44.0.2
ctranslate2==4.5.0
cut-cross-entropy==25.1.1
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dataclasses-json==0.6.7
datasets==3.4.1
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.2.18
diffusers==0.32.2
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
ecdsa==0.19.1
einops==0.8.1
elastic-transport==8.17.1
elasticsearch==8.17.2
encodec==0.1.1
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.12
faster-whisper==1.1.1
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
funcy==2.0
g2pkk==0.1.2
greenlet==3.1.1
griffe==1.7.2
grpcio==1.71.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
hf_transfer==0.1.9
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.5.0
InstructorEmbedding==1.0.1
isodate==0.6.1
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llama-cloud==0.1.17
llama-cloud-services==0.6.9
llama-index==0.12.28
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.28
llama-index-embeddings-huggingface==0.5.2
llama-index-embeddings-instructor==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-huggingface==0.4.2
llama-index-llms-huggingface-api==0.4.1
llama-index-llms-openai==0.3.30
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-elasticsearch==0.4.2
llama-parse==0.6.4.post1
llvmlite==0.43.0
looseversion==1.3.0
lxml==5.4.0
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.8.0.87
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
onnxruntime==1.21.0
packaging==24.2
pandas==1.5.3
passlib==1.7.4
pathlib==1.0.1
peft==0.15.0
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
preshed==3.0.9
propcache==0.3.1
protobuf==3.20.3
prov==2.0.1
psutil==7.0.0
psycopg2-binary==2.9.10
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.5
pynndescent==0.5.13
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
pypinyin==0.54.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytorch-triton==3.3.0+git96316ce5
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
sentence-transformers==2.7.0
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.1
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
starlette==0.46.1
striprtf==0.0.26
SudachiDict-core==20250129
SudachiPy==0.6.10
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
text-generation==0.7.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.8.0.dev20250408+cu128
torchaudio==2.6.0.dev20250408+cu128
torchvision==0.22.0.dev20250408+cu128
tqdm==4.67.1
trainer==0.0.36
traits==7.0.2
transformers==4.50.1
triton==3.2.0
trl==0.15.2
TTS==0.22.0
typeguard==4.4.2
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.13.0
tyro==0.9.17
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
